Change indicator date type of ListHistoriticalAgentReport.

+++ ccc/include/alibabacloud/ccc/model/ListHistoricalAgentReportResult.h
@@ -73,7 +73,7 @@ namespace AlibabaCloud
 							long totalTalkTime;
 							long maxDialingTime;
 							long maxTalkTime;
-							long averageWorkTime;
+							float averageWorkTime;
 							long maxWorkTime;
 							float answerRate;
 							float averageTalkTime;
